Fixes errors in ListViewLoader. Fixes layout problems in EncyclopediaLanding.xml.
