+++ app/src/main/res/layout/activity_encyclopedia_landing.xml
@@ -22,14 +22,16 @@
         android:layout_width="fill_parent"
         android:layout_height="match_parent"
         android:id="@android:id/list"
-        android:layout_weight="1" />
+        android:layout_weight="1"
+        android:layout_below="@+id/button3"
+        android:layout_alignParentStart="true" />
 
 
     <RadioGroup
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignTop="@android:id/list"
-        android:layout_alignEnd="@android:id/list">
+        android:layout_alignParentTop="true"
+        android:layout_alignParentEnd="true">
         <RadioButton
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
